+++ src/lib/mem.h
@@ -6,7 +6,7 @@
 #include "../include/types.h"
 
 void memcpy(u8* source, u8* dest, int no_bytes);
-void memset(u8* dest, u8 val, u32 len);
+void memset(u64* dest, u8 val, u32 len);
 int memcmp(const void* ptr1, const void* ptr2, uint16_t num);
 
 /* At this stage there is no 'free' implemented. */
